Remove blog posts from pages folder
Posts need to be wrapped with a component that can access metadata. This commit moves the posts so that they can be imported as components and wrapped. This can be also be achieved by exporting the wrapper in every MDX file or by passing the wrapper to MDXProvider (no access to metadata).